refactor: move getDateDiff to utils and create fn for lazy state.
